backend/requirements.txt
asgiref==3.5.0
Django==4.0.3
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
pymongo==3.12.1
python-decouple==3.8
pytz==2023.3
sqlparse==0.2.4
typing_extensions==4.6.2
